Organize name accumulator tests
